NCBC-2061: Implement Retry Handling API - Config and K/V only
Motivation
----------
Provide retry handling per RFC #49 for idempotent operations and
transient errors.

Modifications
-------------
 - Move KV execute operations from bucket to the ClusterNode that is the target
 node. The bucket is responsibly for determining which node to send the
 operation to via its KeyMapper.
 - Add an IRetryable marker interface for exceptions that may be
 retried.
 - Add new exceptions to handle specific error cases
 - Add IRetryStrategy interface that represents an interface for a retry
 strategy.
 - Add RetryReasons enumeration and means of mapping between the
 exceptions and the retry reason
 - Add FailFastRetryStrategy and BestEffortRetryStrategy per the RFC
 - Add ICalculateBackoff and implementations for controlled backoff and
 expontential backoff calculators
 - Add Idempotent field to IOperation to flag idempotent operations
 - Add Attempts, RetryReason, and RetryStrategy to IOperation
 - Add RetryOrchestrator that ties everything togather.

result
------

Reviewed-on: http://review.couchbase.org/116551
Tested-by: Build Bot <build@couchbase.com>

+++ src/Couchbase/Core/IO/Operations/Legacy/Errors/RetrySpec.cs
@@ -48,7 +48,7 @@ namespace Couchbase.Core.IO.Operations.Legacy.Errors
         /// </summary>
         /// <param name="attempts">The attempts.</param>
         /// <returns>The next interval to wait before the next</returns>
-        public int GetNextInterval(int attempts)
+        public int GetNextInterval(uint attempts)
         {
             var adjustedAttempts = attempts - 1;
 
@@ -64,7 +64,7 @@ namespace Couchbase.Core.IO.Operations.Legacy.Errors
                     nextInterval += Interval;
                     break;
                 case RetryStrategy.Linear:
-                    nextInterval += adjustedAttempts * Interval;
+                    nextInterval += (int)adjustedAttempts * Interval;
                     break;
                 case RetryStrategy.Exponential:
                     nextInterval += (int) Math.Pow(Interval, adjustedAttempts);
